+++ hf2p_server/source/networking/network_utilities.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void network_get_build_identifiers(int* executable_type, int* executable_version, int* compatible_version);
+void network_get_build_identifiers(long* executable_type, long* executable_version, long* compatible_version);
